Update Shader header

+++ includes/Shader.h
@@ -89,7 +89,7 @@ public:
 	}
 	void setFloat(const std::string &name, float value) const
 	{
-		glUniform1f(glGetUniformLocation(ID, name.c_str()), (int)value);
+		glUniform1f(glGetUniformLocation(ID, name.c_str()), value);
 	}
 	void setVec2(const std::string &name, const glm::vec2 &value) const
 	{
